Throw in fuzzer when the container size is not variable and the requested size cant be achieved due to cap. (#4045)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4045

PR 3887 limited the max size of an elements vector by introducing a cap and
taking the min during the vector constructions.

However, when the user sets   opts.containerVariableLength = false, then he
might be building logic on top of that assumption, for example in one of my diffs
I am testing subscript function and hence generating valid indices based on
the assumption that containerVariableLength is satisfied.

This diff does the following:
1. make default of opts.containerVariableLength to be true.
2. if opts.containerVariableLength=false, and the requested size cant be served
then throw user error.

Alternativly we can make the capping optional and disabled by default.

Reviewed By: pedroerp

Differential Revision: D43324613

fbshipit-source-id: 2ae090cc8408195d24f1fb95fb7ad52f62e0eaa2