Separating grouped and ungrouped drivers. (#4254)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4254

To support mixed execution mode (when some pipelines run in
grouped execution mode and some in ungrouped) in a Task we split
drivers into two separate sets and manage their lifetime separately.
This change does not include the managing of join bridges or local
exchanges, so the mixeed mode execution queries are still failing.

Differential Revision: D43972189

fbshipit-source-id: 5826e8be708db32aef62e55a15e069e312513bde