Do not block LocalPartition operator waiting for peers (#4385)
Summary:
LocalPartition operator doesn't need to be blocked waiting for peers to finish
processing or for LocalExchanges to drain the queues. It can finish as soon as
it pushed all input into the queues.

Modify LocalPartition::close() to not close the queues and let LocalExchange
operators (the consumers) close the queues.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4385

Reviewed By: xiaoxmeng

Differential Revision: D44301161

Pulled By: mbasmanova

fbshipit-source-id: d72858113580bcb515e47e7dbbf1eb583d5ab8b4