Make array_frequency() deterministic (#4435)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4435

Changing how F14 is used to make the output of array_frequency()
deterministic. Since F14 does not provide ordering guarantees, we do another
pass in the input to ensure the output matches the input order (and hence is
deterministic), updating the map frequencies to prevent duplicated values.

Also moving the map to a class member to prevent re-allocations per row.

Reviewed By: laithsakka

Differential Revision: D44425744

fbshipit-source-id: dd573b0dd94b9d57f0fffaa57de75467897268ed