Peel past a loaded lazy layer during expression evaluation (#4547)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4547

This patch enables peeling past a loaded lazy layer. This ensures
that vectors are consistently peeled across instances of shared
sub-expressions. This avoids a bug where partial peeling from a
parent expression can result in removal of lazy layer during
evaluation of children and as a result a shared sub-expressions
can sometimes be executed with inputs peeled till the lazy layer
and sometimes peeled past it.

Differential Revision: D44765830

fbshipit-source-id: fc3dd21c6190878b02adee87e1f9e50d88568a51