Fix crash in Cast when copying localResult smaller than rows (#4511)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4511

Fuzzer found a bug where CastExpr::apply() attempts to copy values in localResult to
result at `rows`. The problem is that localResult is of the size `nonNullRows.end()`
that is smaller than `rows` and hence out-of-bound access happens. This diff fixes
this problem by using `nonNullRows` instead of `rows` when copying localResult.
Since nonNullRows can be empty when the input is an all-null vector, this diff also
updates the condition of adding Nulls to result after the copy.

This diff fixes https://github.com/facebookincubator/velox/issues/4509.

Reviewed By: bikramSingh91

Differential Revision: D44691358

fbshipit-source-id: a2393bf22c432f0a67a9d3f917cb3a313c174041