Safer decoding for ConstantVector<bool> (#4536)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4536

I found a thing that I feel is "dangerous " that we do when we decode constantVector<bool> which does not produce bugs in the current usage.
but its just dangerous.

what we do is that we store a pointer to the bool value_ variable in the
constant vector and refer to that by data_ where its assumed to be
uint64_t* and not bool*
 bits::isBitSet(reinterpret_cast<const uint64_t*>(data_), index(idx))

so if dereferenced it will access that field plus other random memory.
in the case of flatVector it refers to to actual uint64_t*.

we do not have issues right with this because we isBitSet will reference the
first byte only because index would be zero, and we know the first byte is
valid.

also note that we do expose const T* data() const {..} to user and usually with the bool case
T is considered  uint64_t*.

Reviewed By: kgpai, bikramSingh91

Differential Revision: D44743242

fbshipit-source-id: 3e60cb3ad641836568fbae440106f30c08b3ec2b