Set output batch rows adaptively for operators
Add the outputBatchRows() to Operator interface to determine the number
of output batch rows based on the total batch size config
preferred_output_batch_bytes and the given average row size.
If the average row size is not given, return the config
preferred_output_batch_rows.

Make the HashAggregation, OrderBy and Windown operator use the above logic
to adaptively set the number of output batch rows based on the estimated
output row size.