Fix excessive array hash size for sparse tables
An array of up to 16 MB can be allocated for an array based hash table
with only two keys if these kays are about one million apart. This
will cause a partial aggregation hash table to hit the 16 MB default
limit immediately after creation. This in turn will cause it to flush
in the worst case after every batch of input, which will convert a
100% reducing partial agg into no reduction and high overhead.

To fix this, we check, when about to flush a partial aggregation, if
we are in array hash mode with a large array and sparse entries. If
so, we disable range value ids for array based aggregation and
reselect a hash mode. The choice to disable array hash tables for
range value ids stays on for the lifetime of the hash table.