Use estimateFlatSize for dictionary input vector for LocalExchangeQueue
LocalPartition is blocked if the total buffered input size of
the LocalExchangeQueue hits the max_local_exchange_buffer_size.
Since the input is dictionary vector, we should use estimateFlatSize()
instead of retainedSize() for size estimate for the input. retainedSize()
is the size of the wrapped base vector, and thus is ~15 times larger
than the actual size. This makes LocalPartition blocked for every
input vector.