Suppress errors for unused arguments of null-propagating functions (#3189)
Summary:
Suppress errors for unused arguments of null-propagating functions

Null propagating functions will not be called if at least one argument
is null. Also, other arguments for rows where one argument was null
will not be evaluated for the null rows. This may cause errors to be
masked in a way depending on argument order. This makes a commutative
function like + not be commutative in the presence of masked errors.

We defer signalling of errors for arguments of null propagating
functions. An error will be signalled only if there is no null on its
row.

The errors for each Expr subtree are treated separately. Errors from
subtrees to the left must not null out rows for a subtree to the
right. Only arguments of the same function may interact in this way,
hence extra logic for scoping errors in Expr::evalAll.

We divide argument evaluation into distinct functions for default null
behavior and non-null propagating cases.

Expr::evalAll works as before for functions with non-default null
behavior.  default null behavior, evalArgsDefaultNulls first captures
the errors in effect at entry. It then calls eval on each argument
with an empty error state in EvalCtx. It reads the errors from EvalCtx
after each argument and collects these in 'argumentErrors'. If an
argument eval returned null and no error for a row, subsequent
arguments will not be evaluated for this row. After
  all arguments are evaluated, errors are kept only for those rows
 where there is no null with no error.  If errors are left
at this time, they are added to the errors in effect at entry. If we
are in a mode where errors are signalled as they happen
(EvalCtx::throwOnError), we throw.

This change may cause previously reported errors to no longer be
reported or a different error to be reported if there are multiple
errors on a row.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/3189

Reviewed By: bikramSingh91

Differential Revision: D41189273

Pulled By: oerling

fbshipit-source-id: f87c71e01e01ef2f2c27830d327cff3c13677ddc