Bypass non-reducing partial aggregation
If a partial aggregation does not reduce data size even after
increasing its memory, we stop doing it. Adds a toIntermediate method
to GroupingSet and Aggregates. This translates a vector of raw input
into a vector of accumulators, each initialized from the single input
value on the row.

If toIntermediate is not defined for one or more of the aggregates,
GroupingSet::toIntermediate makes temporary rows in the RowContainer
and uses the regular Aggregate methods to make a single-value
accumulator.