Eagerly report completed split groups.
Summary:
We have found that during split execution with 1 concurrent
group (frequent setting) we have 5 seconds spent on a single split
group processing.
This is due to message roundtrip:
1. Splits come for group A
2. Drivers are created for group A and start running
3. "No more splits for group A" message comes (can be on step 1).
4. ALL drivers finished -> We mark split group A as finished.
5. Coordinator polls us and we send "split group A finished" info to it.
6. Splits come for group B
7. And so on.

The change here is to report a completed split group as soon as
we have received all splits for it.
Spits for queued split groups are simply waiting for their turn to be run.
The change does not affect what splits groups the worker get - there is
no competition between workers for split groups.
The result is the execution time reduction from 3 minutes to 25 seconds
on 60 worker cluster.

Differential Revision: D45360901

fbshipit-source-id: b3755e93502cd5402dc5e589a3eabe1b474ba34e