Fix NOT handling in metadata filter (#4756)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4756

If a condition is evaluated to true for a row group, it means there are
some rows in the group that can pass the condition.  Negating the condition
should mean whether there are some rows that does not pass the condition.  When
some rows pass the condition and some others not in the group, we can not simply
negate the result for a inner node.  We must push the NOT down to leaf node and
combine it with leaf filter in this case.

Fix https://github.com/facebookincubator/velox/issues/4749

Reviewed By: oerling

Differential Revision: D45324976

fbshipit-source-id: 0afbec4237bf486311c68d0a1ac03f89d8d70194