Readers/RowReaders must own their 'options_'
A Reader/RowReader reader  must own its Reader/RowReaderOptions.
Parquet Reader and RowReader and the DWRF RowReader were erroneously holding a reference to Options owned by the containing DataSource.

The reference will point to freed memory in the event of async inited
RowReader being used on a different DataSource. A Reader and RowReader
must be movable between DataSources, as is done in
setFromDataSource().

The error is demonstrated by e.g. velox_tpch_benchmark. The error has
previously not surfaced because after initialization the options were
typically not accessed. With the addition of mutability there are
checks that start failing.