Start prefetch and coalesce of large colunnbs earlier
A column spanning multiple load quanta is coalescable to surrounding
columns if it qualifies for prefetch. Like this, the cache entries of
all the quanta are filled from one IO.

The test for prefetch compares references to reads. However the
current CachedBufferedInput has recorded the reference but not the
read since the read is only future. So, a fully read large column
comes out wiht a read density of 50% on the second time unless the
reference added by the current load is subtracted.

With multiple threads the references will still outpace the reads for the first few row groups, e.g. The column has been referenced 30 times for 15 reads with 15 reading threads. On the third row group there will be 45 references and 30 reads.

Therefore we also drop the minimum read density (% of referenced data that is accessed) to 60 from 80.