Fix batch of all inactive conditional aggregates
In GroupingSet::toIntermediate, when there is a batch of all masked
out inputs for an aggregate, the result is not a column of all nulls
in all cases. For example, count has a column of all zeros. Therefore,
if all inputs are masked out, initialize the aggregate on one row and
read this post-initialization state to all elements of the result
vector. The result vector is always a preallocated, reused flat
vector, so there is no point with constant wrapping or the like for a
rare special case.