Add feature projection visibility metrics (#4831)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4831

To know when feature projection is failing or succeeding, we want to provide some data from reader level on how many keys we are filtering with flatmap.

This:
- Creates a callback from the Koski layer which adds to aggregator used by koski QueryStats, and stores it to dwio config
- Passes callback from dwio config to row reader options
- Passes callback from row reader options to flatmap context
- Passes callback from row reader options t
- Aggregates total and selected key streams in flatmap column reader. The delta between these reflects filtering done by feature projection
- Fires callback in flatmap column reader once filtering is done

The reason for so many handoffs is because dwio is not designed to have config available in column reader level. Exposing metrics from this level to koski user is a bit of a hack with our current metrics framework.

This is behind a JustKnobs flag:

https://www.internalfb.com/intern/justknobs/?name=xldb_ml_engine%2Fkoski#enable_feature_projection_metrics

Reviewed By: Magoja

Differential Revision: D45508059

fbshipit-source-id: 08bb48f05c1bca8530c9c08f9b430484863e99c6