Add MemConfigMutable class. (#4839)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4839

MemConfigMutable acts just like MemConfig, but allows
modifying existing properties and adding new ones.
It also protects the storage with folly::Synchronized.

It is planned to be used in Presto Native, where we, based on
a single property will either create MemConfig (in production)
or MemConfigMutable (in test and staging).

The goals is to allow changing some properties on the fly w/o
restarting/pushing pkg to the clusters.

Reviewed By: gggrace14

Differential Revision: D45580035

fbshipit-source-id: 0515883dbb26458fa7c319eb11e9e549067bda20