Enhance fuzzer to generate multiple expression trees per step (#4703)
Summary:
This adds a feature in expression fuzzer that allows us to generate
multiple expressions in each step that would be executed in the same
ExprSet. This allows us to add coverage for cases where common sub
expressions and column references exist between expression trees
and verify that optimizations that are employed in these cases
work as expected.
Note: This changes how result vectors are generated and serialized
for repro however it ensures that the ExpressionRunner still remains
compatible with the previous format so that previous repro files work.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/4703

Reviewed By: kagamiori

Differential Revision: D45282488

Pulled By: bikramSingh91

fbshipit-source-id: 9237aec356eeb253741fc111d00d7194a2559c69