Add fuzzer support for feeding preloaded lazy vectors
This change adds the ability to randomly pick columns in the input
row vector to wrap in lazy and load before feeding to common
evaluation path. It helps to simulate the case where a lazy vector
is loaded in a previous step before being fed to the evaluator.
(It would help catch issues like the one fixed in #4547)
It re-purposed the mechanism already in place that generates a list
of column indices to wrap in lazy and enhances it to use negative
indices as well which represent columns indices that should be
preloaded. This makes the change backwards compatible with repro
files generated before it and ensures minimal code changes.