Fix decoding of percentiles in ApproxPercentileAggregate::addIntermediate
Summary:
ApproxPercentileAggregate::addIntermediate() receives an input vector of Row type and SelectivityVector `rows` for this input vector. addIntermediate() currently uses `rows` to decode a child of the base vector of the input vector.

This diff fixes it by translating `rows` to the selected base rows before decoding the child vector.

Differential Revision: https://internalfb.com/D45751685

fbshipit-source-id: c2093b7dc65552731835906a25dd4f343fd5554f