Fix decoding of percentiles in ApproxPercentileAggregate::addIntermediate (#4903)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4903

ApproxPercentileAggregate::addIntermediate() receives an input vector of Row type and SelectivityVector `rows` for this input vector. addIntermediate() currently uses `rows` to decode a child of the base vector of the input vector.

This diff fixes it by translating `rows` to the selected base rows before decoding the child vector.

Reviewed By: Yuhta

Differential Revision: D45751685

fbshipit-source-id: dc6841e37f8bfc1fce93769b9fdbc8e97fab4d88