Symplifying QueryConfig. (#4950)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4950

Main changes:
- Removing unused classes ConfigStack, BaseConfigManager & Context.
- Making QueryConfig an owner of its config.

presto-trunk build will fail - for this I will need to create a small PR in GH (https://github.com/prestodb/presto/pull/19668), merge it and wait till diff train brings it to fbcode, then rebase this diff and it would succeed.
We can proceed with reviewing w/o waiting for all of this.

Reviewed By: xiaoxmeng

Differential Revision: D45898645

fbshipit-source-id: ac7b69059f22bba81d113a50acbe7057cd66330e