Don't merge regions for vread (#5010)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5010

For wsVRead we need to pass all regions by separate.

`enqueuedToBufferOffset_` was holding an intermediate value between sort and merge. But now, since we may not merge, se save the intermediate result in `offsetToEnqueued_`.

I also realized that, if we don't merge, we don't need to create another `regions` variable for calling vread. We can just use `regions_`.

Differential Revision: D46012935

fbshipit-source-id: 450d706218f1fdd6ba7c7a4ecd89a1e55ed96cc9