Fix ElementAt unit test with no input vectors (#4788)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4788

There are two unit test of ElementAt that evaluates on expressions
that only consist of literal values with no input vector. The test utility
method infers the selectivity vector for evaluation to be of length 0,
while the test case actually expect to evaluate on one single row.
This diff extends the `evaluate` and `evaluateOnce` APIs to accept
an optional SelectivityVector and fix the ElementAt unit tests.

Reviewed By: laithsakka

Differential Revision: D45420524

fbshipit-source-id: 7322892b3160a60a662279af2cfe72aea4c51f49