Add fast path to SubscriptUtil if the elements are fixed width (#4991)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/4991

Today, SubscriptUtil always creates a Vector around the Array elements as a dictionary and constructs the indices into the elements.

If the elements are fixed width, the cost of constructing the indices is no more than the cost of copying the elements into a FlatVector.  Constructing the FlatVector saves on decoding the DictionaryVector downstream (we have seen cases where this savings is substantial).

If the elements are themselves encoded (rather than flat) this does incur the additional cost of decoding.  Since we're probably going to pay that cost eventually anyway (unless the query filters out all the rows the Subscript is applied to), I did not limit this to the case where the elements are flat.

Reviewed By: laithsakka

Differential Revision: D45972924

fbshipit-source-id: 4de59c2a59178a9ab19904ac6d7c1bd56b895945