Differentiate taskCompletionFuture and stateChangeFuture.
Summary:
stateChangeFuture was returning futures realized when the Task is finished,
not when the state was changed, per se.
We've been using this method to wait for Task state change when replying to
getTaskStatus() and getTaskInfo() messages.
This created the delay for up to 1 second in the Tasks with grouped execution,
when we completed splits, but report them only when 1 second timeout in
the message reply code is expired, not immediately.
This change introduces the new taskCompletionFuture method that realizes
futures when the Task is completed, like stateChangeFuture did before.
stateChangeFuture on the other hand now realizes futures upon the Task
completion as well as upon a split group completion.
This improves the grouped execution query timing in the following way:
  1st run (clear cache): 1:22 - 1:27 -> 0:59
  Subsequent runs (warm cache): 0:38 - 0:42 -> 0:16 - 0:29

Also refactored TaskCompletionNotifier by moving it from the h to cpp and renaming.

Changed unit tests and benchmarks to use the new taskCompletionFuture.

Differential Revision: D46411874

fbshipit-source-id: 40b3b96f9302f7766ca12696a2992dac4a6c83bc