Disallow registering aggregation functions with names already exist (#5080)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5080

The current implementation of registerAggregateFunction() silently
overwrites the function registry if a user attempts to register a function
with a name already exsits. This diff makes registerAggregateFunction()
not overwrite the function registry and return false in this situation. User
can pass an boolean flag to force overwriting if necessary.

Besides, the registration function of some aggregation functions always return true, while others return the return value of exec::registerAggregateFunction. This diff unify them to all return the return value of exec::registerAggregateFunction.

Reviewed By: Yuhta

Differential Revision: D46306052

fbshipit-source-id: e7882f9dca5ea59d6a82c6e5545838716887ac33