Optimize SelectiveFlatMapColumnReader memory footprint
Summary: `SelectiveFlatMapColumnReader` is having memory issue when we read data that has many columns (e.g. 20000).  Instead of keeping all child readers in memory, we create one single reader at a time and drop the reader after the reading.  In next read we create a new child reader and seek to the right row group and offset for the child reader.

Differential Revision: D46489132

fbshipit-source-id: cf3e25bfb02b6de76d789b37017740692d11ee80