Fix FieldReference changing input vector when it's multiply referenced (#5079)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5079

Fuzzer found a bug in FieldReference. When accessing a constant input vector, FieldReference resizes the original vector if the VectorPtr is unique. But for an expression `if(c0, c1, c1)`, c1's VectorPtr is unique, but c1 is multiply referenced. Resizing c1 to a shorter size when evaluating the then-branch causes incorrect result because the else-branch won't be able to copy enough values from c1 to the result vector.

Reviewed By: laithsakka

Differential Revision: D46243192

fbshipit-source-id: 99933ef3972c79372600cc4bee961d6e32798a52