Revert D46494517: Multisect successfully blamed D46494517 for test or build failures
Summary:
This diff is reverting D46494517
D46494517: [velox] Add fallthrough to variant::toJson by Yuhta has been identified to be causing the following test or build failures:

Tests affected:
- [hpc/torchrec/models/feed/gpu_tests:ifr_combined_mtml_launcher_test - test_packaging (hpc.torchrec.models.feed.gpu_tests.ifr_combined_mtml_launcher_test.IFRCombinedMTMLLauncherTest)](https://www.internalfb.com/intern/test/562950043789773/)

Here's the Multisect link:
https://www.internalfb.com/multisect/2217013
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D46558239

fbshipit-source-id: 5564c4f80b10a826f9583dec2b5f14fc5a751348