Share filterCache memory when the dictionary is shared (#5205)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5205

The integer dictionary used by flatmap values are usually shared, but in our code we are instantiating different filter cache for them and wasted memory as well as CPU to reallocate and recalculate same data.  This change makes the filter cache shared among child readers of flatmap if the dictionary is shared.

Differential Revision: D46560720

fbshipit-source-id: 63956711a19cc3b533bb5a616538847159c5a16f