Fix decodedVector's state when re-used (#5286)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5286

Expression evaluation re-used decoded vectors as an optimization. A decoded vector has a member variable that stores a vector of indices merged across all encoding layers while decoding a vector. This is currently only resized when used so if a decodedVector ends up being re-used with a different set of selected rows, then the unselected rows might have indices from a previous usage. This can be a problem for instance where nulls and indices are generated after decoding and do not have access to the original set of selected rows which were used for decoding, because it tried to create a copy and can end up using old/ garbage indices that can result in a segfault (see nulls()).
This fix makes sure that the internal member variable keeping track of indices is reset to a default value whenever it is used.

Reviewed By: Yuhta

Differential Revision: D46767207

fbshipit-source-id: 4a169bd625b09d2750f7bc868f2f61a5bb3a2dcd