Fix hangs in MultiFragmentTest
Summary:
We've been seeing intermittent hangs in
MultiFragmentTest.raceBetweenTaskTerminateAndTaskNoMoreSplits, e.g.
https://app.circleci.com/pipelines/github/facebookincubator/velox/27164/workflows/b3f6d810-9ea7-4e0c-9513-c187a66cbe5e/jobs/171592

I managed to reproduce locally and found it was stuck waiting on the EventCount
blockTerminate.

Looking at the documentation for the EventCount class, we aren't using it in the
recommended way.  In particular, it's subject to the spurious wake ups and races
that condition variables are, and we need to handle that.
https://github.com/facebook/folly/blob/main/folly/experimental/EventCount.h#L33

Fortunately, there's an await method, which combined with an atomic variable
makes it easy.  I've updated all the EventCount objects in this test and the rest of
the file to use this approach to avoid the race condition we were hitting.

Differential Revision: D46907187

fbshipit-source-id: 6b60dd5bb185ff5d986b6ddb22f4eca55657403d