Add capability to read from coalesced IOBuf into region IOBufs (#5327)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5327

In the new preadv model we'll return a vector of IOBufs, so we don't reserve memory before the call (since our current network protocol returns IOBufs). This way the caller will get the IOBufs and reserve the memory, if necessary, after the call, so the duplication (if any) is short lived.

Our current coalescing logic returns one IOBuf for multiple Regions. We have to split it into multiple IOBufs. They'll just wrap the same underlying buffer as the main coalesced read IOBuf.

Reviewed By: Yuhta

Differential Revision: D46843005

fbshipit-source-id: 2047f27d73abc6a3acaf4bb9c97516c1a3ced5d6