Fix race condition in memory accounting in LocalPartition's LocalExchangeQueue
Summary:
The test LocalPartitionTest.blockingOnLocalExchangeQueue has been flaky.

I've traced this to a race condition in how memory is tracked in
LocalExchangeQueue's LocalExchangeMemoryManager.

In enqueue we lock the queue, put the item on, unlock the queue, then try to mark
the memory as acquired in LocalExchangeMemoryManager via
increaseMemoryUsage.

In next, we lock the queue, remove the item, mark the memory as released in
LocalExchangeMemoryManager via decreaseMemoryUsage, then unlock the
queue. (It's the same in close, but that's not relevant here).

This can lead to negative memory usage, e.g. the following operations can be
interleaved as so:
* Put the item on the queue
* Take the item off the queue and mark the memory as freed (negative)
* Mark the memory as acquired (back to 0)

Worse we could block the wrong thread:
* Put small item1 on the queue
* Put big item2 on the queue
* Take item1 off the queue and mark its memory as freed
* Mark the memory for item2 as acquired (hitting the max memory)
* Mark the memory for item1 as acquired, see that the max has now been hit and
mark the thread blocked on consumer even though item1 isn't on the queue
anymore

Calling increaseMemoryUsage with the lock on the queue held when we put the
item on the queue fixes this race condition.  Note that the lock acquired in
increaseMemoryUsage is only otherwise acquired in decreaseMemoryUsage which
is only called with the lock on the queue held, so the queue lock and the memory
lock are acquired in a consistent order meaning this doesn't introduce deadlocks.

Differential Revision: D46909393

fbshipit-source-id: 1fbe9497804a8cbb24a4e916884e817a06225fa2